Styles Re-factorying People Page & global.ts - Planets Page, PlanetsTable & Planet Badge Components, Routing, and Dynamic Rendering
App:
Following the same structure of People Component Routing Logic, implemented the Planet Component with dynamic rendering, Started the development of the PlanetsPage component only displaying and fetching the start of the data structure of the api.

My focus today was trimming up the project, applying the re-utilizable styles to global, to better maintain and read the code.

Added some quotes to better organize the file global.ts and removed some style.ts that we're no longer being used.