People Page Refactor
App:
Added a default SVG to deal with some endpoints that are at the moment empty.

People Page:
Still need to treat the loading to adapt to all dynamic data that are consumed by the new functions and components.
Refactor the whole page to adapt to better scalability and readability, all sections are now components, aligned with the default styles for them.

Components Added:
+ Character
+ HomeWorld
+Species
+Films(static)
+StarShips
+Vehicles

Except from the Films Component, the rest is receiving and consuming data from their respective endpoints.

Vehicles & StarShips:
Since the data fetched was sometimes and array of two strings for the requests to the respective data, i had to develop a logic to adapt the data fetched and request at all iterations of the data fetched.

For that i used a for loop with axios.all and maping the strings after i stored them to a var (urlStrings) and then using axios.get to retrieve the data for each iteration of the response.
Then i rendered the badges for Vehicles and  StarShips with the map method, to display the number of badges to match the data call.