Planets Page (Logic Complete), People Page : Films Component (Complete) & FilmPage Characters Component Added
App:

Planets Page;
All logic for the planet page is functional, have to deal with a few bugs but the main structure is complete.
Films Component is iterating dynamically from the endpoint and map rendering the results, the image is still static since the api doesn't provide one.

Added:
+ FIlms Component

PeoplePage:
Also added logic to the Films component to iterate through the end point data results. and all logic for PeoplePage is done and has the same issues of PlanetsPage that need to be dealt with, and if species,starships, vehicles badge changes whatsoever the styles should be updated if make sense.

FilmPage:
Characters component added using similar logic and styles to residents component in PlanetPage component

Added:
+Characters Component