People Page - Films Component new useEffect structure, and types refactor
People Page:
Refactor the PeoplePageProps type, creating DataProps type to better readability of the code.

Films Component:
Refactor the functional component to a single useEffect. Using modern practices to improve the readability and scalability of the code.

Homeworld Component:
Created a verification to render and fetch the data from the endpoint when the PeoplePage is rendered.

Next step is adapting the new functional structure to the components that follow the same logic as the Films Component, this structure was a little complex to adapt the data structure of the swapi.