Functional Page, Modern Structure & Types
As all Pages components are now structured to the new defined structure, they're all working as intended.
Only have to adapt the react Type bug when rendering a .map, and also typing the arr from all of the table components. Then will be good to make a production build and deploy the site via netlify.